Add DeltaTransaction trait and update post-commit hooks with it
