+++ spark/src/main/scala/org/apache/spark/sql/delta/hooks/UpdateCatalog.scala
@@ -26,7 +26,7 @@ import scala.util.control.NonFatal
 
 import org.apache.spark.sql.delta.skipping.clustering.{ClusteredTableUtils, ClusteringColumnInfo}
 import org.apache.spark.sql.delta.skipping.clustering.temp.ClusterBySpec
-import org.apache.spark.sql.delta.{DeltaConfigs, DeltaTableIdentifier, OptimisticTransactionImpl, Snapshot}
+import org.apache.spark.sql.delta.{DeltaConfigs, DeltaTableIdentifier, DeltaTransaction, Snapshot}
 import org.apache.spark.sql.delta.actions.{Action, Metadata}
 import org.apache.spark.sql.delta.logging.DeltaLogKeys
 import org.apache.spark.sql.delta.metering.DeltaLogging
@@ -64,7 +64,7 @@ trait UpdateCatalogBase extends PostCommitHook with DeltaLogging {
 
   override def run(
       spark: SparkSession,
-      txn: OptimisticTransactionImpl,
+      txn: DeltaTransaction,
       committedVersion: Long,
       postCommitSnapshot: Snapshot,
       actions: Seq[Action]): Unit = {
